refactor: 对齐变量值

+++ models/message.go
@@ -8,7 +8,7 @@ import (
 // Message 消息模型
 type Message struct {
 	MessageID    int64     `json:"message_id" gorm:"primaryKey"`
-	SenderUserID int64     `json:"sender_user_id" validate:"required"`
+	SenderUserID uint      `json:"sender_user_id" validate:"required"`
 	ChannelID    string    `json:"channel_id" validate:"required"`
 	Content      string    `json:"content" validate:"required"`
 	Attachment   string    `json:"attachment"`
